Django_tests/django_test_05-12-22/requirements.txt
asgiref==3.5.0
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
django-filter==22.1
drf-yasg==1.21.4